Rejigger project structure
