+++ NdnGameBackend/build.gradle
@@ -3,7 +3,7 @@
  */
 
 dependencies {
-    compile project(':NdnCore')
+    compile project(':NdnGameCore')
     compile 'com.fasterxml.jackson.core:jackson-databind:2.9.8'
     compile 'com.google.inject.extensions:guice-assistedinject:3.0'
 }
