+++ NdnGameCore/src/main/java/com/stefanolupo/ndngame/Player.java
