+++ NdnGameCore/src/main/java/com/stefanolupo/ndngame/config/ConfigIF.java
