+++ NdnGameCore/src/main/java/com/stefanolupo/ndngame/exceptions/NdnException.java
