+++ NdnGameCore/src/main/java/com/stefanolupo/ndngame/names/BaseName.java
