+++ NdnGameCore/src/main/java/com/stefanolupo/ndngame/names/BulletsName.java
