+++ NdnGameCore/src/main/java/com/stefanolupo/ndngame/names/PlayerStatusName.java
