+++ NdnGameCore/src/main/proto/NdnGame.proto
