+++ NdnGameCore/src/main/resources/log4j_config.properties
