--- NdnGameLibGdx/core/build.gradle
@@ -1,11 +1,0 @@
-apply plugin: "java"
-
-sourceCompatibility = 1.6
-[compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
-
-sourceSets.main.java.srcDirs = [ "src/" ]
-
-
-eclipse.project {
-    name = appName + "-core"
-}
