--- NdnGameLibGdx/desktop/src/com/stefnaolupo/ndngame/libgdx/desktop/DesktopLauncher.java
@@ -1,12 +1,0 @@
-package com.stefnaolupo.ndngame.libgdx.desktop;
-
-import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
-import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
-import com.stefnaolupo.ndngame.libgdx.LibGdxGame;
-
-public class DesktopLauncher {
-	public static void main (String[] arg) {
-		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-		new LwjglApplication(new LibGdxGame(), config);
-	}
-}
