+++ NdnGameLibGdxCore/assets/badlogic.jpg
