+++ NdnGameLibGdxCore/build.gradle
@@ -1,0 +1,6 @@
+apply plugin: "java"
+
+sourceCompatibility = 1.8
+[compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
+
+//sourceSets.main.java.srcDirs = [ "src/" ]
