+++ NdnGameLibGdxCore/src/Game.gwt.xml
