+++ NdnGameLibGdxCore/src/main/java/com/stefanolupo/ndngame/libgdx/LibGdxGame.java
@@ -1,0 +1,27 @@
+package com.stefanolupo.ndngame.libgdx;
+
+import com.badlogic.gdx.Game;
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+
+public class LibGdxGame extends Game {
+	private SpriteBatch batch;
+	private MainScreen mainScreen;
+
+	@Override
+	public void create () {
+		batch = new SpriteBatch();
+		mainScreen = new MainScreen(this);
+		setScreen(mainScreen);
+	}
+
+	@Override
+	public void render () {
+		super.render();
+	}
+	
+	@Override
+	public void dispose () {
+		batch.dispose();
+		screen.dispose();
+	}
+}
