+++ NdnGameLibGdxDesktop/src/main/java/com/stefanolupo/ndngame/libgdx/desktop/DesktopLauncher.java
@@ -1,0 +1,12 @@
+package com.stefanolupo.ndngame.libgdx.desktop;
+
+import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
+import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
+import com.stefanolupo.ndngame.libgdx.LibGdxGame;
+
+public class DesktopLauncher {
+	public static void main (String[] arg) {
+		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
+		new LwjglApplication(new LibGdxGame(), config);
+	}
+}
